# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /users:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    get:
      description: Get all users
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UsersResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Create a user
      parameters:
        - in: body
          name: user
          description: The user to create.
          schema:
            type: object
            required:
              - username
              - password
              - firstname
              - lastname
            properties:
              username:
                type: string
              password:
                type: string
              roles:
                type: array
                items:
                  type: string
                  enum: [user, admin]
              firstname:
                type: string
              lastname:
                type: string
      responses:
        "201":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UsersResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  # /users/create:
  #   # binds a127 app logic to a route
  #   x-swagger-router-controller: user
    
  /users/{username}:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    parameters:
        - name: username
          in: path
          description: username of the user to get
          required: true
          type: string
    get:
      description: Get user by username
      # used as the method name of the controller
      operationId: get
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UsersResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{username}/edit:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    parameters:
        - name: username
          in: path
          description: username of the user to get
          required: true
          type: string
    post:
      description: Get user by username
      # used as the method name of the controller
      operationId: edit
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UsersResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{username}/delete:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    parameters:
        - name: username
          in: path
          description: username of the user to get
          required: true
          type: string
    post:
      description: Get user by username
      # used as the method name of the controller
      operationId: destroy
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UsersResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  UsersResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string